Dialog_FieldDefinition: Don't offer a self-triggering lookup.
Document: Add get_relationships_excluding_triggered_by().
Dialog_FieldDefinition::set_field(): Use it to avoid offering
a relationship in the list of lookup relationships if the
relationship has the actual field as its from field, because
that would be a self-triggering relationship (endless).

Ideally, we would have some way to show the user why they don't
see the relationship in the list, but at least we will prevent
them from doing this.

The previous commit checks for this too.

Bug #754641 (m.rick.mac)