no raw loops: Add string_join() and use it instead of for loops.
Inspired by Sean Parent's "no raw loops" idea:
 https://www.youtube.com/watch?v=qH6sSOr-yk8
though I am not really convinced that this improves the code in this
particular case. What we gain by having no for loop we lose by having the
awkard lambda syntax. In Sean Parent's examples this is clearer because of
the use of (I think) adaptors to refer to particular member fields.