Build: Set GLOM_LOCALEDIR from localdir, as set by configure.
Like generated po/Makefile.in.in file does, to decide where to
instal the translation files.
Instead of guessing it based on the prefix, causing us not to show
the translations when staring in appropriate locale.
Bug #761372 (m.rick.mac)
And this fixes the 3 test cases that have been failing for me for
a while.