Box_Tables::on_adddel_add(): Don't allow a space in a table name.
Warn the user if the table name hase a space and then don't try to
create the table. This is not user friendly but I would prefer not to
escape/unescape all the table names. That seems fragile and I would
rather hope that the libgda bug is fixed one day.
Bug #763534 (m.rick.mac).

Also adjust the test because we currently expect this to fail.