libglom: Copy in the libepc source code.
Intead of using it as a shared library. This avoids the need
for distro to package it, though its only used by Glom now,
and avoids the need for me to maintain the libepc-ui code against
the changing GTK+ 4 API, though Glom doesn't use libepc-ui.

This does not include the libepc-ui code from libepc.